+++ cg/server/api.py
@@ -47,7 +47,7 @@ from cg.server.dto.orders.order_delivery_update_request import OrderDeliveredUpd
 from cg.server.dto.orders.order_patch_request import OrderDeliveredPatch
 from cg.server.dto.orders.orders_request import OrdersRequest
 from cg.server.dto.orders.orders_response import Order, OrdersResponse
-from cg.server.dto.sequencing_metrics.sequencing_metrics import SequencingMetricsRequest
+from cg.server.dto.sequencing_metrics.sequencing_metrics_request import SequencingMetricsRequest
 from cg.server.ext import db, delivery_message_service, lims, order_service, osticket
 from cg.store.models import (
     Analysis,
